pyproject.toml
[project]
name = "daily-digest"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "feedparser>=6.0.11",
    "google-api-python-client>=2.110.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "schedule>=1.2.1",
    "youtube-transcript-api>=0.6.2",
    "markdown>=3.5",
]
